+++ demos/11-bezier-perpendiculars/index.js
@@ -28,7 +28,7 @@
 		      scaleX                 : 1.0,
 		      scaleY                 : 1.0,
 		      rasterGrid             : true,
-		      drawOrigin             : true,
+		      drawOrigin             : false,
 		      rasterAdjustFactor     : 2.0,
 		      redrawOnResize         : true,
 		      defaultCanvasWidth     : 1024,